refactor /nettverking
