+++ public/sitemap-0.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://itxbergen.no/frivillig</loc><lastmod>2025-08-28T17:22:38.288Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://itxbergen.no/styremedlem</loc><lastmod>2025-08-28T17:22:38.289Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://itxbergen.no/karrieredag</loc><lastmod>2025-08-28T17:22:38.289Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://itxbergen.no/karrieredag/studenter</loc><lastmod>2025-08-28T17:22:38.289Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://itxbergen.no/icon.ico</loc><lastmod>2025-08-28T17:22:38.289Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://itxbergen.no/nettverking</loc><lastmod>2025-08-28T17:22:38.289Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://itxbergen.no/sitemap.xml</loc><lastmod>2025-08-28T17:22:38.289Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://itxbergen.no/kontakt</loc><lastmod>2025-08-28T17:22:38.289Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://itxbergen.no/personvern</loc><lastmod>2025-08-28T17:22:38.289Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://itxbergen.no/karakterkalkulator</loc><lastmod>2025-08-28T17:22:38.289Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://itxbergen.no/karrieredag/bedrifter</loc><lastmod>2025-08-28T17:22:38.289Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://itxbergen.no/om-itx</loc><lastmod>2025-08-28T17:22:38.289Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://itxbergen.no/stillingsannonser</loc><lastmod>2025-08-28T17:22:38.289Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://itxbergen.no</loc><lastmod>2025-08-28T17:22:38.289Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no/frivillig</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no/karakterkalkulator</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no/styremedlem</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no/karrieredag/bedrifter</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no/karrieredag</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no/karrieredag/studenter</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no/kontakt</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no/nettverking</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no/icon.ico</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no/personvern</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no/stillingsannonser</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no/sitemap.xml</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no/om-itx</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://itxbergen.no</loc><lastmod>2025-08-29T17:30:16.851Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
 </urlset>
