+++ src/app/nettverking/page.tsx
@@ -1,11 +1,19 @@
+import Otherthings from "@/components/utils/otherthings";
+import PageHeader from "./page-header";
+import SectionStudentOrganisations from "./section-student-organisations";
 import { Metadata } from "next";
-import Nettverking from "./nettverking";
 
 export const metadata: Metadata = {
   title: "Nettverking | ITxBergen",
   description: "Nettverking for IT-studenter i Bergen",
-}
+};
 
-export default function NettverkingPage() {
-  return <Nettverking />;
+export default function Nettverking() {
+  return (
+    <div>
+      <PageHeader />
+      <SectionStudentOrganisations />
+      <Otherthings />
+    </div>
+  );
 }
