+++ src/lib/student-organisations.ts
@@ -1,0 +1,204 @@
+export const studentOrganisations = [
+  {
+    name: "echo",
+    location: "Universitetet i Bergen",
+    boxHeader:
+      "Linjeforeningen er et sosialt og faglig bindeledd mellom studentene og ledelsen ved Institutt for informatikk, UiB",
+    boxp1:
+      "Linjeforeningen jobber for å bedre det sosiale miljøet til studentene, samt ta opp faglige saker. Foreningen er også ett bindeledd mellom studentene og andre studentpolitiske organisasjoner ved Universitetet i Bergen. I tillegg til å kommunisere med instituttet, gjør echo mange små og store tiltak for studentene. Dette inkluderer alt fra å starte opp studentpub til å ha filmklubb, eller å dele ut pris til studentene som er flinkest til å holde lesesalen ryddig.",
+    website: "https://echo.uib.no/",
+    instagram: "https://www.instagram.com/echo_uib/",
+    facebook: "https://www.facebook.com/groups/informatikk/",
+    linkedin:
+      "https://www.linkedin.com/company/echo-linjeforeningen-for-informatikk/",
+    participants: "Informatikkstudenter er automatisk medlem av echo",
+    address: "Institutt for Informatikk, Thormøhlensgate 55, 5008 Bergen",
+    epost: "echo@uib.no",
+  },
+  {
+    name: "Root",
+    location: "Høgskulen på Vestlandet",
+    boxHeader:
+      "Root Linjeforening er en studentorganisasjon ved Høgskulen på Vestlandet for Data- og IT-studenter.",
+    boxp1:
+      "Organisasjonen arrangerer blant annet sosiale arrangementer og arbeider for å knytte studentene nærmere næringslivet gjennom bedriftspresentasjoner og workshops. Arrangementer blir organisert gjennom skoleåret, med formål om å skape bekjentskaper på tvers av klassetrinn og fremme et godt studentmiljø ved instituttet. Programmet varierer fra temafester til quizer, og samarbeid har også blitt inngått med andre organisasjoner.",
+    website: "https://rootlinjeforening.no/",
+    instagram: "https://www.instagram.com/linjeforeningenroot/",
+    facebook: "https://www.facebook.com/RootLinjeforening",
+    linkedin: "https://www.linkedin.com/company/root-linjeforening/",
+    participants: "Alle data/IT-studenter ved HVL i Bergen.",
+    epost: "linjeforening.root@hvl.no",
+    address: "Inndalsveien 28, 5063 Bergen",
+  },
+  {
+    name: "Cloud",
+    location: "Høyskolen Kristiania",
+    boxHeader:
+      "Vi er Cloud, linjeforeningen for alle IT-studenter ved Høyskolen Kristiania campus Bergen.",
+    boxp1:
+      "Cloud har siden IT-linjene startet opp ved Campus Bergen i 2020 jobbet for å gjøre studiehverdagen til medstudentene våre bedre, i tillegg til å fungere som et bindeledd mellom studentene og næringslivet.",
+    boxp2:
+      "Vårt ønske er å la våre medstudenter se hvor mange muligheter som finnes innen IT-verden da vår erfaring er at mange studenter ikke vet hvor mange bedrifter som driver med IT eller har egen IT-avdeling. Dette gjør vi gjennom å arrangere både sosiale arrangement og bedriftspresentasjoner med næringslivet i Bergen og omegn.",
+    website: "https://linktr.ee/cloud_linjeforening",
+    instagram: "https://www.instagram.com/linjeforeningencloud/",
+    facebook: "https://www.facebook.com/LinjeforeningenCloud",
+    linkedin: "https://www.linkedin.com/company/cloud-linjeforeningen-for-it/",
+    participants:
+      "Alle som studerer IT-bachelor på Kristiania Campus Bergen. Campus Bergen tilbyr per nå to IT-bachelorgrader: Cybersikkerhet og Informasjonsteknologi: Frontend- og mobilutvikling.",
+    epost: "leder.cloud@studentunionen.no",
+    address: "Kalfarveien 78C, 5022 Bergen",
+  },
+  {
+    name: "Cyber",
+    location: "Høgskulen på Vestlandet",
+    boxHeader:
+      "Cyber er linjeforeningen for studentene på Cyberfysisk Nettverksteknologi ved Høgskulen på Vestlandet (HVL).",
+    boxp1:
+      "Vårt mål er å styrke både det faglige og sosiale samholdet blant studentene, samtidig som vi legger til rette for kontakt med næringslivet gjennom bedriftsbesøk og samarbeid",
+    website: "https://cyberstudent.no/",
+    instagram: "https://www.instagram.com/linjeforeningencyber/",
+    linkedin: "https://www.linkedin.com/company/cyber-linjeforening/",
+    participants:
+      "Alle studenter på Cyberfysisk Nettverksteknologi ved HVL i Bergen.",
+    epost: "leder@cyberstudent.no",
+    address: "Inndalsveien 28, 5063 Bergen",
+  },
+  {
+    name: "CybVest",
+    location: "Høyskolen Kristiania",
+    boxHeader:
+      "Vi er en linjeforening hovedsakelig for Cyber Studenter, men også alle IT studenter her på Høyskolen Kristiania.",
+    boxp1:
+      "Vi jobber for å skape spennende sosiale- og faglige arrangementer for studentene. Alle våre arrangementer er gratis å delta på, så mest mulig folk kan være med! Vi planlegger blant annet bedriftspresentasjoner og spillkvelder hvert semester.",
+    website: "https://www.studentunionen.no/",
+    instagram: "https://www.instagram.com/cybvest/",
+    facebook: "https://www.facebook.com/profile.php?id=61552474341943",
+    linkedin: "https://www.linkedin.com/company/cybvest-linjeforening/about/",
+    participants:
+      "Hovedsakelig så er det IT / Cyber studenter på HK, men vi er absolutt åpen for at andre IT studenter kan komme på arrangementene våre hvis de vil!",
+    epost: "leder.cybvest@studentunionen.no",
+    address: "Kalfarveien 78C, 5022 Bergen",
+  },
+  {
+    name: "Enigma",
+    location: "Universitetet i Bergen",
+    boxHeader:
+      "Enigma er linjeforeningen for alle studentene ved sivilingeniørprogrammet Informasjonsteknologi og økonomi (ITØK) ved Universitetet i Bergen.",
+    boxp1:
+      "Formålet med linjeforeningen er å bidra til et godt læringsmiljø blant studentene, legge til rette for sosiale aktiviteter og fremme kontakten med næringslivet. I tillegg skal foreningen være med å støtte opp om studentenes interesser og krav ovenfor instituttet samt andre relevante organ. Alle studentene på ITØK regnes som medlemmer av Enigma, og foreningen ledes av styret bestående av åtte medlemmer.",
+    boxp2: "Enigma har flere undergrupper og flere på vei, blant annet: ",
+    list: [
+      "Fintech Enigma",
+      "Enigma Badeklubb",
+      "Enigma Friluftsliv",
+      "Enigma Bryggelag",
+    ],
+    website: "https://enigma-web.onrender.com",
+    instagram: "https://www.instagram.com/enigma.uib/",
+    facebook: "https://www.facebook.com/profile.php?id=61573198165416",
+    linkedin: "https://www.linkedin.com/company/enigmaitok/",
+    participants:
+      "Alle studenter ved sivilingeniørprogrammet Informasjonsteknologi og økonomi (ITØK) ved UIB.",
+    address: "Fosswinckels Gate 14, 5007 Bergen",
+    epost: "Enigma.forening@gmail.com",
+  },
+  {
+    name: "Enter",
+    location: "Universitetet i Bergen",
+    boxHeader:
+      "Vi i Enter er studentforeningen ved infomedia-instituttet på Universitetet i Bergen.",
+    boxp1:
+      "Det vil si en felles organisasjon for hele instituttet vårt og ikke bare en enkelt linje. Arrangementene vi holder er både sosiale og faglige. De sosiale arrangementene kan være alt fra quiz på diskuterbar eller grilling i parken til gokart eller egne fester på utesteder. De faglige arrangementene kan være bedriftspresentasjoner med ulike bedrifter eller workshops. Innad i Enter har vi en egen bedriftskomite, sosialkomité og PR. Her er det egne oppdrag for hver av komiteene, men vi samarbeider også mye på tvers av hverandre.",
+    website: "https://www.svstudent.no/enter",
+    instagram: "https://www.instagram.com/enter_uib/",
+    linkedin:
+      "https://www.linkedin.com/company/studentforeningen-enter/posts/?feedView=all",
+    participants: "Infomedia-studenter er automatisk medlem av Enter",
+    address:
+      "Det Samfunnsvitenskapelige Fakultet, Fosswinckels gate 6, Bergen 5007",
+    epost: "Enter@uib.no",
+  },
+  {
+    name: "Helt Sikker",
+    location: "Universitetet i Bergen",
+    boxHeader:
+      "Er du en student med interesse for datasikkerhet? Er du keen på å lære mer om sikkerhet, og få solid erfaring med styreverv og drifting av en studentorganisasjon?",
+    boxp1:
+      "Helt Sikker er en studentorganisasjon ved Universitetet i Bergen. Vår målsetning er å tilby studenter en arena hvor man kan lære mer om informasjonssikkerhet, samtidig som vi bidrar til å bygge et miljø for sikkerhetsinteresserte studenter i Bergen.",
+    boxp2:
+      "Helt Sikker ble originalt stiftet av studenter som tilhørte det første kullet av UiBs bachelorprogram i datasikkerhet. Et av våre mål er å komplementere UiBs mer teoretiske tilnærming til sikkerhet ved å arrangere workshops, foredrag, CTF-er, konkurranser og mer.",
+    website: "https://heltsikker.no/",
+    instagram: "https://www.instagram.com/heltsikker_bergen/",
+    facebook: "https://www.facebook.com/groups/heltsikker/",
+    linkedin: "https://www.linkedin.com/company/helt-sikker/?viewAsMember=true",
+    participants: "Studenter ved UiB interessert i datasikkerhet",
+    epost: "styret@heltsikker.no",
+    address: "Ingen adresse",
+  },
+  {
+    name: "Inkogvito",
+    location: "Universitetet i Bergen",
+    boxHeader:
+      "Inkogvito sitt hovedfokus er å styrke det sosiale samholdet og det faglige innholdet knyttet til studiet vårt.",
+    boxp1:
+      "Kognitiv Vitenskap er et lite studie hvor hvert kull kun består av 20-30 studenter hvert år. Når klassene er såpass små er det sosiale samholdet og kulturen til studentene desto mer viktig. Derfor ble Inkogvito stiftet i våren 2021 med mål om å bedre studiehverdagen til alle våre studenter. Med dette nye sentraliserte organet for Kognitiv Vitenskap ble det også mulig å ha et mer overordnet fokus på næringsliv og den faglige strukturen til studiet vårt.",
+    instagram: "https://www.instagram.com/inkogvito.uib/",
+    facebook: "https://www.facebook.com/groups/inkogvito",
+    linkedin: "https://www.linkedin.com/company/inkogvito/",
+    participants:
+      "Alle som studerer Kognitiv Vitenskap har allerede åpen invitasjon til å engasjere seg. Du trenger ikke registrere deg for å bli med.",
+    epost: "inkogvito@uib.no",
+    address: "Ingen adresse",
+  },
+  {
+    name: "ReMIX",
+    location: "Universitetet i Bergen",
+    boxHeader:
+      "ReMIX er linjeforeningen for medie- og interaksjonsdesign ved UiB.",
+    boxp1:
+      "Vi jobber for å skape et inkluderende miljø som styrker både det sosiale og faglige. Gjennom arrangementer, workshops og samarbeid med bransjen kobler vi studiet til arbeidslivet, samtidig som vi bygger et sterkt studentfellesskap.",
+    instagram: "https://www.instagram.com/remix_uib/",
+    facebook: "https://www.facebook.com/groups/535661775706870/",
+    participants:
+      "Alle studenter ved medie- og interaksjonsdesign ved UiB kan være med i linjeforeningen. (master og bachelor)",
+    epost: "remix@uibstudent.no",
+    address: "Media City Bergen",
+  },
+  {
+    name: "friByte",
+    location: "Annet",
+    boxHeader:
+      "friByte er en studentorganisasjon som har sitt utspring fra Det Akademiske Kvarters tidligere IT-avdeling, E-tjenesten.",
+    boxp1:
+      "friByte er en null-profitt organisasjon, som innebærer at vi ikke tar utbytte av overskuddet, og at medlemmene ikke mottar økonomisk kompensasjon for deres deltagelse. friByte driver nå på sitt 22. år og har nå vokst til å inkludere et utvalg dedikerte studenter og profesjonelle kunder.",
+    boxp2: "Hva er målet til friByte?",
+    list: [
+      "Levere sikre IT-tjenester av høy kvalitet til studentorganisasjoner.",
+      "Opprettholde et aktivt læremiljø for IT-interesserte studenter.",
+      "Pådriver for etisk bruk av programvare og persondata.",
+      "Være en bidragsyter for utvikling og forbedring av åpen og fri programvare.",
+    ],
+    website: "https://fribyte.no/",
+    instagram: "https://www.instagram.com/fribyteno/",
+    facebook: "https://www.facebook.com/profile.php?id=100077992657390",
+    linkedin: "https://www.linkedin.com/company/fribyte/",
+    participants: "Alle våre medlemmer er eller har vært studenter. ",
+    epost: "post@fribyte.no",
+    address: "Ingen adresse",
+  },
+  {
+    name: "E-tjenesten",
+    location: "Annet",
+    boxHeader: "E-tjenesten er IT-gruppen på Det Akademiske Kvarter.",
+    boxp1:
+      "E-tjenesten har ansvar for drift og utvikling av digitale løsninger. Gruppen møtes jevnlig både for arbeid og sosiale aktiviteter. Det er lav terskel for å bli med, og man trenger ingen forkunnskaper. Som medlem får man tilgang til interne fordeler og blir del av et sosialt og faglig miljø.",
+    website: "https://kvarteret.no/e-tjenesten-info/",
+    instagram: "https://www.instagram.com/kvarteretbergen/",
+    facebook: "https://www.facebook.com/@Kvarteret/?locale=nb_NO",
+    linkedin: "https://no.linkedin.com/company/det-akademiske-kvarter",
+    participants:
+      "Alle studenter som er nysgjerrige på IT og har lyst til å lære mer, kan bli med i E-tjenesten. Du trenger ingen forkunnskaper – bare interesse og motivasjon! ",
+    epost: "it.leder@kvarteret.no",
+    address: "Olav Kyrres gate 49, 5015 Bergen",
+  },
+];
