+++ src/types/student-organisation.ts
@@ -1,0 +1,16 @@
+export interface StudentOrganisation {
+  name: string;
+  location: string;
+  boxHeader: string;
+  boxp1: string;
+  boxp2?: string;
+  list?: string[];
+  address?: string;
+  participants?: string;
+  epost?: string;
+  website?: string;
+  facebook?: string;
+  instagram?: string;
+  linkedin?: string;
+  github?: string;
+}
